% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vital.R
\name{load_vital}
\alias{load_vital}
\title{Load Vital}
\usage{
load_vital(path, tz = "UTC")
}
\arguments{
\item{path}{Path to folder containing the files exported by utilities/vital_s3.exe}

\item{tz}{Timezone used for converting Unix epochs to datetime.}
}
\value{
A (nested) list of tracks.
}
\description{
Loads a folder of gz compressed csv files, exported by the vital utility vital_s3.exe
}
\details{
Before import, a .vital file must be converted to csv files using the untility program vital_s3.exe.
This can be found in \verb{<vitalrecorder-folder>/utilities}.

From the command line:
\verb{cd <path-to-vitalrecorder-folder>/utilities}
\verb{vital_s3.exe <input.vital> <output-folder>}
}
\examples{
test_folder <- system.file('extdata', 'test_data_demo', package = 'vitalrecordR')
load_vital(test_folder, tz = 'CET')
}
