% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vital.R
\name{load_vital_header}
\alias{load_vital_header}
\title{Load Header CSV from S3 export}
\usage{
load_vital_header(path, tz)
}
\arguments{
\item{path}{Path to folder containing the files exported by utilities/vital_s3.exe}

\item{tz}{Timezone used for converting Unix epochs to datetime.}
}
\description{
Load Header CSV from S3 export
}
