% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vital.R
\name{read_vital_header}
\alias{read_vital_header}
\title{Read Header CSV from S3 export}
\usage{
read_vital_header(path, tz)
}
\arguments{
\item{path}{Path to folder containing the files exported by utilities/vital_s3.exe}

\item{tz}{Timezone used for converting Unix epochs to datetime.}
}
\description{
Read Header CSV from S3 export
}
